+++ CHANGES.md
@@ -54,6 +54,17 @@ Changed:
     - all arguments are now also available as alternative long names
 - io.TextMenu: `width` is now an optional parameter. If not defined, surface width is based on the widest item in the menu
 - misc.get_system_info: format changes for ``as_text`` output and improvements
+- audiosystem:
+    - support for selecting audio device
+    - new control default ``audiosystem_device``
+    - new function ``misc.get_audio_devices``
+- stimuli.Video:
+    - general improvements
+    - Pygame video backend removed (always relies on mediadecoder)
+    - Pygame audio backend is now the default (instead of sounddevice)
+    - new parameter ``audio_backend``
+    - new stimuli default ``video_audio_backend``
+    - Pygame audio backend uses current audiosystem (if started)
 
 Fixed:
 - bug in colour.is_hex & colour.is_colour